LayoutTests/storage/domstorage/documentURI-expected.txt

Test that changing documentURI has no effects on storage events.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Testing sessionStorage
storage.clear()
PASS storage.length is 0
window.onstorage = firstEvent
storage.foo = '123'
First event fired
Saving URI
document.documentURI = 'abc'
PASS document.documentURI is "abc"
window.onstorage = secondEvent
storage.foo = 'xyz'
Second event fired
PASS true is true


Testing localStorage
storage.clear()
PASS storage.length is 0
window.onstorage = firstEvent
storage.foo = '123'
First event fired
Saving URI
document.documentURI = 'abc'
PASS document.documentURI is "abc"
window.onstorage = secondEvent
storage.foo = 'xyz'
Second event fired
PASS true is true
PASS successfullyParsed is true

TEST COMPLETE
